+++ _unittests/ut_serverdoc/test_ping.py
@@ -59,7 +59,7 @@ class TestPing(unittest.TestCase):
             self._testMethodName,
             OutputPrint=__name__ == "__main__")
 
-        if is_travis_or_appveyor() in ("appveyor", "travis"):
+        if is_travis_or_appveyor():
             return
         machine = self.get_machine()
         out = regular_ping_machine(machine, delay=0.1, nb_max=3, fLOG=fLOG)