+++ src/pyquickhelper/loghelper/run_cmd.py
@@ -271,7 +271,7 @@ def run_cmd(cmd, sin="", shell=sys.platform.startswith("win"), wait=False, log_e
                     "communicate should be True to send something on stdin")
             stdout, stderr = pproc.stdout, pproc.stderr
 
-            begin = time.clock()
+            begin = time.perf_counter()
             last_update = begin
             # with threads
             (stdoutReader, stdoutQueue) = _AsyncLineReader.getForFd(
@@ -289,7 +289,7 @@ def run_cmd(cmd, sin="", shell=sys.platform.startswith("win"), wait=False, log_e
                     if fLOG is not None:
                         fLOG(prefix_log + sdecol)
                     out.append(sdecol)
-                    last_update = time.clock()
+                    last_update = time.perf_counter()
                     if stop_running_if is not None and stop_running_if(decol, None):
                         runloop = False
                         break
@@ -302,18 +302,18 @@ def run_cmd(cmd, sin="", shell=sys.platform.startswith("win"), wait=False, log_e
                     if fLOG is not None:
                         fLOG(prefix_log + sdecol)
                     err.append(sdecol)
-                    last_update = time.clock()
+                    last_update = time.perf_counter()
                     if stop_running_if is not None and stop_running_if(None, decol):
                         runloop = False
                         break
                 time.sleep(0.05)
 
-                delta = time.clock() - last_update
+                delta = time.perf_counter() - last_update
                 if tell_if_no_output is not None and delta >= tell_if_no_output:
                     fLOG(prefix_log + "[run_cmd] No update in {0} seconds for cmd: {1}".format(
                         "%5.1f" % (last_update - begin), cmd))
-                    last_update = time.clock()
-                full_delta = time.clock() - begin
+                    last_update = time.perf_counter()
+                full_delta = time.perf_counter() - begin
                 if timeout is not None and full_delta > timeout:
                     runloop = False
                     fLOG(prefix_log + "[run_cmd] Timeout after {0} seconds for cmd: {1}".format(